pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nudibranch"
version = "0.0.1"
description = "Terminal dashboard for worldwide dive conditions monitoring"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Nudibranch Team"}
]
dependencies = [
    "httpx>=0.25.0",
    "aiocache>=0.12.2",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "tenacity>=8.2.3",
    "diskcache>=5.6.3",
    "python-dotenv>=1.0.0",
    "textual>=0.47.0",
    "rich>=13.7.0",
    "pyTMD>=2.1.0",
    "numpy>=1.24.0",
    "copernicusmarine>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-httpx>=0.27.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.1",
    "ruff>=0.1.6",
]

[project.scripts]
nudibranch = "nudibranch.tui.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"
